add: リファクタリング
