+++ PositiveWordsCollection/Firebase/DataService.swift
@@ -17,139 +17,10 @@ class DataService {
     private let userCollection = Firestore.firestore().collection("users")
     @AppStorage(CurrentUserDefaults.userID) var currentUserID: String?
 
-    func postDelete(postID: String) async throws {
-            // SubCollection Delete
-            await subCollectionDelete(postID: postID)
-            // PostCollection Delete
-            try await Firestore.firestore().collection("posts").document(postID).delete()
-            // Storage削除
-            await postsStorageDelete(postID: postID)
-    }
-    // サブコレクションの削除が完了した後に親ドキュメントも削除する
-    func deleteAccount(userID: String) async throws {
-        // Authアカウント削除
-        try await AuthService.instance.userAcountDelete()
-        // UserDefault削除
-        // All UserDefault Delete
-        let defaultDictionary = UserDefaults.standard.dictionaryRepresentation()
-        defaultDictionary.keys.forEach { key in
-            UserDefaults.standard.removeObject(forKey: key)
-        }
-        await deleteUserCollection(userID: userID)
-        // posts Collection of userID
-        try await postAllDelete(userID: userID)
-        // Storage削除
-        await userStorageDelete(userID: userID)
-        // users Collection Delete
-        try await userCollection.document(userID).delete()
-    }
-
-    private func postAllDelete(userID: String) async throws {
-        // posts Collection of userID
-            let postOfUserSnapshot =  try await Firestore.firestore().collection("posts").whereField(DatabasePostField.userID, isEqualTo: userID).getDocuments()
-
-        for document in postOfUserSnapshot.documents {
-            let postID = document.documentID
-            // SubCollection Delete
-            await subCollectionDelete(postID: postID)
-            print("⭐️これPostsDoCument\(document.documentID)⭐️")
-            // PostCollection Delete
-            try await Firestore.firestore().collection("posts").document(postID).delete()
-            // Storage削除
-            await postsStorageDelete(postID: postID)
-        }
-    }
-
-    private func subCollectionDelete(postID: String) async {
-        let subCollection = Firestore.firestore().collection("posts").document(postID).collection("comments")
-        do {
-            let subSnapshot = try await subCollection.getDocuments()
-            for subdocument in subSnapshot.documents {
-                try await subdocument.reference.delete()
-            }
-        } catch {
-            print("subCollectionDelete Error")
-        }
-    }
-
-    private func postsStorageDelete(postID: String) async {
-        let storageRef = Storage.storage().reference()
-        let postIDRef = storageRef.child("posts").child(postID).child("1")
-        do {
-            try await postIDRef.delete()
-        } catch {
-            print("postsStorageDelete Error")
-        }
-    }
-
-    private func userStorageDelete(userID: String) async {
-        let storageRef = Storage.storage().reference()
-        let userIDRef = storageRef.child("users").child(userID).child("profile")
-        do {
-            try await userIDRef.delete()
-        } catch {
-            print("userStorageDelete Error")
-        }
-    }
-
-    private func deleteUserCollection(userID: String) async {
-        do {
-            let usersAccountSnapshot = try await Firestore.firestore().collection("users").whereField(DatabasePostField.userID, isEqualTo: userID).getDocuments()
-            for usersDocument in usersAccountSnapshot.documents {
-                try await usersDocument.reference.delete()
-            }
-        } catch {
-            print("deleteUserCollection Error")
-        }
-    }
-
-    func uploadPost(image: UIImage, caption: String, displayName: String, userID: String) async {
-        // Create new post document
-        let document = postsREF.document()
-        let postID = document.documentID
-        // Upload image to Storage
-        do {
-            try await ImageManager.instance.uploadPostImage(postID: postID, image: image)
-            let postData = [
-                DatabasePostField.postID: postID,
-                DatabasePostField.userID: userID,
-                DatabasePostField.displayName: displayName,
-                DatabasePostField.caption: caption,
-                DatabasePostField.dateCreated: FieldValue.serverTimestamp()
-            ] as [String: Any]
-            try await document.setData(postData)
-        } catch {
-            print("Error uploading post image to firebase")
-        }
-    }
-    // 🟥報告
-    func uploadReport(postID: String) async throws {
-        let data: [String: Any] = [
-            DatabaseReportsField.postID: postID,
-            DatabaseReportsField.dateCreated: FieldValue.serverTimestamp()
-        ]
-        try await reportsREF.addDocument(data: data)
-    }
-
-    // commentsSubCollection
-    func uploadComment(postID: String, content: String, displayName: String, userID: String) async throws -> String? {
-        let document = postsREF.document(postID).collection(DatabasePostField.comments).document()
-        let commentID = document.documentID
-        let data: [String: Any] = [
-            DatabaseCommentsField.commentID: commentID,
-            DatabaseCommentsField.userID: userID,
-            DatabaseCommentsField.content: content,
-            DatabaseCommentsField.displayName: displayName,
-            DatabaseCommentsField.dateCreated: FieldValue.serverTimestamp()
-        ]
-        try await document.setData(data)
-        return commentID
-    }
     // MARK: Get functions
     // UserIDの投稿を取得
     func downloadPostForUser(userID: String) async throws -> [PostModel] {
         let querySnapshot = try await postsREF.whereField(DatabasePostField.userID, isEqualTo: userID).getDocuments()
-        let docData = querySnapshot.documents
         return try await getPostsFromSnapshot(querySnapshot: querySnapshot)
     }
 
@@ -158,19 +29,6 @@ class DataService {
         // 最新の50個しか取得しない
         let querySnapshot = try await postsREF.order(by: DatabasePostField.dateCreated, descending: true).limit(to: 50).getDocuments()
         return try await getPostsFromSnapshot(querySnapshot: querySnapshot)
-    }
-    //　❤️
-    func myLiked(postID: String, userID: String) async throws -> Bool {
-        let query = postsREF.document(postID).collection(DatabasePostField.likedBy).whereField(DatabaseLikedByField.userID, isEqualTo: userID)
-        let countQuery = query.count
-                let myLikeCountSnapshot = try await countQuery.getAggregation(source: .server)
-                print(myLikeCountSnapshot.count)
-                let count = myLikeCountSnapshot.count as? Int ?? 0
-                if count >= 1 {
-                    return true
-                } else {
-                    return false
-                }
     }
 
     private func getPostsFromSnapshot(querySnapshot: QuerySnapshot?) async throws -> [PostModel] {
@@ -188,7 +46,7 @@ class DataService {
                     var likeByUser: Bool = false
                     // ❤️自分がいいねを押したか？UserID
                     if let userID = currentUserID {
-                     likeByUser = try await DataService.instance.myLiked(postID: postID, userID: userID)
+                        likeByUser = try await DataService.instance.myLiked(postID: postID, userID: userID)
                     }
                     // NewPost
                     let newPost = PostModel(postID: postID, userID: userID, username: displayName, caption: caption, dateCreated: date, likeCount: likeCount, likedByUser: likeByUser, comentsCount: commentCount)
@@ -229,25 +87,60 @@ class DataService {
     }
 
     // MARK: UPDATE FUNCTION
-    // 💛
-    func commentCount(postID: String) async throws -> Int {
-        let query = postsREF.document(postID).collection(DatabasePostField.comments)
-        let countQuery = query.count
-        let snapshot = try await countQuery.getAggregation(source: .server)
-        print("❤️\(snapshot.count)❤️")
-        return snapshot.count as? Int ?? 0
+    func uploadPost(image: UIImage, caption: String, displayName: String, userID: String) async {
+        // Create new post document
+        let document = postsREF.document()
+        let postID = document.documentID
+        // Upload image to Storage
+        do {
+            try await ImageManager.instance.uploadPostImage(postID: postID, image: image)
+            let postData = [
+                DatabasePostField.postID: postID,
+                DatabasePostField.userID: userID,
+                DatabasePostField.displayName: displayName,
+                DatabasePostField.caption: caption,
+                DatabasePostField.dateCreated: FieldValue.serverTimestamp()
+            ] as [String: Any]
+            try await document.setData(postData)
+        } catch {
+            print("Error uploading post image to firebase")
+        }
     }
-    // 🟥
-//    func commentPostCount(postID: String, currentUserID: String) async throws {
-//        let commentArray = try await downloadComments(postID: postID)
-//        let increment: Int64 = 1
-//        let data: [String: Any] = [
-//            DatabasePostField.commentCount: FieldValue.increment(increment)
-//        ]
-//        try await postsREF.document(postID).updateData(data)
-//    }
-//
-
+    // 🟥報告
+    func uploadReport(postID: String) async throws {
+        let data: [String: Any] = [
+            DatabaseReportsField.postID: postID,
+            DatabaseReportsField.dateCreated: FieldValue.serverTimestamp()
+        ]
+        try await reportsREF.addDocument(data: data)
+    }
+    // commentsSubCollection
+    func uploadComment(postID: String, content: String, displayName: String, userID: String) async throws -> String? {
+        let document = postsREF.document(postID).collection(DatabasePostField.comments).document()
+        let commentID = document.documentID
+        let data: [String: Any] = [
+            DatabaseCommentsField.commentID: commentID,
+            DatabaseCommentsField.userID: userID,
+            DatabaseCommentsField.content: content,
+            DatabaseCommentsField.displayName: displayName,
+            DatabaseCommentsField.dateCreated: FieldValue.serverTimestamp()
+        ]
+        try await document.setData(data)
+        return commentID
+    }
+    //　❤️
+    func myLiked(postID: String, userID: String) async throws -> Bool {
+        let query = postsREF.document(postID).collection(DatabasePostField.likedBy).whereField(DatabaseLikedByField.userID, isEqualTo: userID)
+        let countQuery = query.count
+        let myLikeCountSnapshot = try await countQuery.getAggregation(source: .server)
+        print(myLikeCountSnapshot.count)
+        let count = myLikeCountSnapshot.count as? Int ?? 0
+        if count >= 1 {
+            return true
+        } else {
+            return false
+        }
+    }
     // 💛
     func likeCount(postID: String) async throws -> Int {
         let query = postsREF.document(postID).collection(DatabasePostField.likedBy)
@@ -256,59 +149,31 @@ class DataService {
         print("🩵\(snapshot.count)❤️")
         return snapshot.count as? Int ?? 0
     }
-    //🟡
-//    func unLikeCount(postID: String, myUserID: String) async throws -> Int {
-//        let query = postsREF.document(postID).collection(DatabasePostField.likedBy).whereField(DatabaseLikedByField.userID, isEqualTo: myUserID)
-//        let snapShot = try await query.getDocuments()
-//        for document in snapShot.documents {
-//            try await document.reference.delete()
-//        }
-//        let count = try await likeCount(postID: postID)
-//        return count
-//    }
-    //🩵
-//    func likePost(postID: String, currentUserID: String) {
-//        // Update post count
-//        // Update who liked
-//        let increment: Int64 = 1
-//        let data: [String: Any] = [
-//            DatabasePostField.likeCount: FieldValue.increment(increment),
-//            DatabasePostField.likedBy: FieldValue.arrayUnion([currentUserID])
-//        ]
-//        postsREF.document(postID).updateData(data)
-//    }
+    // 💛
     func unLikePost(postID: String, myUserID: String) async throws {
         let query = postsREF.document(postID).collection(DatabasePostField.likedBy).whereField(DatabaseLikedByField.userID, isEqualTo: myUserID)
         let snapShot = try await query.getDocuments()
         for document in snapShot.documents {
             try await document.reference.delete()
         }
-//        let count = try await likeCount(postID: postID)
-//        return count
     }
-    // ❤️liked_bySubCollection
-//    func uploadLikedPost(postID: String, userID: String) async throws -> String {
+    // 💛
     func uploadLikedPost(postID: String, userID: String) async throws {
         let document = postsREF.document(postID).collection(DatabasePostField.likedBy).document()
-//        let likedByID = document.documentID
         let data: [String: Any] = [
-//            DatabaseLikedByField.likedBy: likedByID,
             DatabaseLikedByField.userID: userID,
-//            DatabaseLikedByField.displayName: displayName,
             DatabaseLikedByField.dateCreated: FieldValue.serverTimestamp()
         ]
         try await document.setData(data)
-//        return likedByID
     }
-//🩵
-//    func unlikePost(postID: String, currentUserID: String) {
-//        let decrement: Int64 = -1
-//        let data: [String: Any] = [
-//            DatabasePostField.likeCount: FieldValue.increment(decrement),
-//            DatabasePostField.likedBy: FieldValue.arrayRemove([currentUserID])
-//        ]
-//        postsREF.document(postID).updateData(data)
-//    }
+    func commentCount(postID: String) async throws -> Int {
+        let query = postsREF.document(postID).collection(DatabasePostField.comments)
+        let countQuery = query.count
+        let snapshot = try await countQuery.getAggregation(source: .server)
+        print("❤️\(snapshot.count)❤️")
+        return snapshot.count as? Int ?? 0
+    }
+
     // MARK: UPDATE USER FUNCTION
     func updateDisplayNameOnPosts(userID: String, displayName: String) async throws {
         let posts = try await downloadPostForUser(userID: userID)
