+++ PositiveWordsCollection/Helper/DatabaseLikedByField.swift
@@ -8,8 +8,6 @@
 import Foundation
 
 struct DatabaseLikedByField {
-//    static let likedByID = "liked_by_id"// 🟥
     static let userID = "user_id"
-//    static let displayName = "display_name"// 🟥
     static let dateCreated = "date_created"
 }
