+++ PositiveWordsCollection/View/SettingsView.swift
@@ -96,7 +96,7 @@ struct SettingsView: View {
                 Task {
                     do {
                         guard let userID = currentUserID else { return }
-                        try await DataService.instance.deleteAccount(userID: userID)
+                        try await DeleteService.instance.deleteAccount(userID: userID)
                         viewModel.showSignInView = true
                     } catch {
                         showDeleteAccountError = true
